==> step_1c_gather_ticker_prices.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
from IPython.display import Markdown as md
from ipynb.fs.full import functions
import pandas as pd

InteractiveShell.ast_node_interactivity = "all"

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

timeline = pd.read_excel('timeline.xlsx', sheet_name='timeline')

tickers = pd.read_excel('tickers.xlsx', sheet_name='tickers')
ticker_list = tickers.copy()['ticker'].to_list()

market_suffix = '.JK'
ticker_list = [f'{ticker}{market_suffix}' for ticker in ticker_list]

pro_or_con_column_name = 'pro_or_con'
pro_or_con_choice = ['PRO', 'CON']

output_sheet_list = []

for pro_or_con in pro_or_con_choice:
    selected_pro_or_con = timeline[timeline[pro_or_con_column_name] == pro_or_con]
    for _, row in selected_pro_or_con.iterrows():
        case_code = row['code']
        start_date = row['scraping_start_date']
        end_date = row['scraping_end_date']
        result = functions.get_ticker_prices(tickers=ticker_list, start_date=start_date, end_date=end_date)
        output_sheet_list.append({
            'sheet_name': case_code,
            'data_frame': result,
        })

with pd.ExcelWriter('ticker_prices.xlsx') as excel_writer:
    for i in output_sheet_list:
        sheet_name = i['sheet_name']
        df = i['data_frame']
        df.index = df.index.map(str)
        df.to_excel(excel_writer=excel_writer, sheet_name=sheet_name)
