==> step_2b_formation_t-test.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
from IPython.display import Markdown as md
from ipynb.fs.full import functions
import pandas as pd

InteractiveShell.ast_node_interactivity = "all"

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

import numpy as np
from scipy.stats import ks_1samp, norm

formation_filename = 'output_formation.xlsx'
formation_excel_book = pd.read_excel(io=formation_filename, header=[0,1], index_col=0, sheet_name=None)

def symbolize(p_value: float) -> str:
  if p_value <= .001:
    return '***'
  elif p_value <= .01:
    return '**'
  elif p_value <= .05:
    return '*'
  else:
    return ''

input_t_test_dict = {
    'PRO': {
        'winner': [],
        'loser': [],
    },
    'CON': {
        'winner': [],
        'loser': [],
    },
}

winner_loser_list = []

for case in formation_excel_book.keys():
    pro_or_con = case.split(sep='-')[0]

    formation_df = formation_excel_book[case]

    factor = int(len(formation_df.columns.levels[0]) / 5)
    ticker_list = formation_df.columns.unique(level=0).to_list()
    winner_ticker_list = ticker_list[:factor]
    loser_ticker_list = ticker_list[-factor:]

    winner_loser_list.append({
        'case': case,
        'winner': ', '.join(winner_ticker_list),
        'loser': ', '.join(loser_ticker_list),
    })

    winner_car = formation_df.copy()[[(i, 'param3_cu') for i in winner_ticker_list]].mean(axis=1).iloc[-1]
    loser_car = formation_df.copy()[[(i, 'param3_cu') for i in loser_ticker_list]].mean(axis=1).iloc[-1]

    input_t_test_dict[pro_or_con]['winner'].append(winner_car)
    input_t_test_dict[pro_or_con]['loser'].append(loser_car)
    
formation_t_test_list = []
for pro_or_con in input_t_test_dict.keys():
    for winner_or_loser in input_t_test_dict[pro_or_con].keys():
        car_list = input_t_test_dict[pro_or_con][winner_or_loser]
        acar = np.mean(car_list)
        t_stat, p_value = ks_1samp(x=car_list, cdf=norm.cdf)
        symbol = symbolize(p_value=p_value)
        formation_t_test_list.append({
            'pro_or_con': pro_or_con,
            'winner_or_loser': winner_or_loser,
            'acar': acar,
            't_stat': t_stat,
            'p_value': p_value,
            'symbol': symbol,
        })

winner_loser_df = pd.DataFrame(winner_loser_list)
formation_t_test_df = pd.DataFrame(formation_t_test_list)

md('**Winner-Loser List**')
winner_loser_df

md('**Formation T-Test Result**')
formation_t_test_df

with pd.ExcelWriter('output_formation_winner-loser_ticker_list.xlsx') as excel_writer:
    winner_loser_df.to_excel(excel_writer=excel_writer, index=None)

with pd.ExcelWriter('output_formation_t-test_result.xlsx') as excel_writer:
    formation_t_test_df.to_excel(excel_writer=excel_writer, index=None)
