==> step_4c_ensemble_gathered_information.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
from IPython.display import Markdown as md
import pandas as pd

InteractiveShell.ast_node_interactivity = "all"

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

import numpy as np

ticker_excel_book = pd.read_excel('tickers.xlsx', sheet_name=0, index_col=0)

winner_loser_ticker_df = pd.read_excel('output_formation_winner-loser_ticker_list.xlsx', sheet_name=0, index_col=0)

u_after_event_excel_book = pd.read_excel(io='gathered_u_after_event.xlsx', sheet_name=None, index_col=0)

u_event_excel_book = pd.read_excel(io='gathered_u_event.xlsx', sheet_name=None, index_col=0)

leak_excel_book = pd.read_excel(io='gathered_leak.xlsx', sheet_name=None, index_col=0)

vlm_excel_book = pd.read_excel(io='gathered_vlm.xlsx', sheet_name=None, index_col=0)

mv_excel_book = pd.read_excel(io='gathered_mv.xlsx', sheet_name=None, index_col=0)

ret_excel_book = pd.read_excel(io='gathered_ret.xlsx', sheet_name=None, index_col=0)

vlt_excel_book = pd.read_excel(io='gathered_vlt.xlsx', sheet_name=None, index_col=0)

cor_df = pd.read_excel(io='gathered_cor.xlsx', sheet_name=0, index_col=0)


case_list = u_after_event_excel_book.copy().keys()
ticker_df = ticker_excel_book.copy()['sector']
cor_series = cor_df.copy().iloc[:,0]
period_first = 1
period_last = 20
period_range = range(period_first, period_last+1)

regression_data_list = []
regression_data_winner_list = []
regression_data_loser_list = []

for case in case_list:
    pro_or_con = case.split('-')[0]
    devent = int(pro_or_con == 'PRO')

    winner_loser_series = winner_loser_ticker_df.loc[case]
    winner_list = winner_loser_series['winner'].split(', ')
    loser_list = winner_loser_series['loser'].split(', ')

    u_0_series = u_event_excel_book[case].copy().iloc[:, 0]
    leak_series = leak_excel_book[case].copy().iloc[:,0]
    ret_series = ret_excel_book[case].copy().iloc[:,0]
    vlt_series = vlt_excel_book[case].copy().iloc[:,0]
    
    for ticker, sector in ticker_df.items():
        u_0 = u_0_series[ticker]
        leak = leak_series[ticker]
        ret = ret_series[ticker]
        vlt = vlt_series[ticker]
        cor = cor_series[sector]

        u_after_event_df = u_after_event_excel_book[case]
        vlm_df = vlm_excel_book[case]
        mv_df = mv_excel_book[case]

        for period in period_range:
            u = u_after_event_df.loc[period, ticker]
            vlm = vlm_df.loc[period, ticker]
            mv = mv_df.loc[period, ticker]

            row = {
                'CASE': case,
                'PRO_OR_CON': pro_or_con,
                'PERIOD': period,
                'TICKER': ticker,
                'U': u,
                'U_0': u_0,
                'LEAK': leak,
                'VLM': vlm,
                'VLM_transformed': np.log(vlm + 1),
                'MV': mv,
                'MV_transformed': np.log(mv),
                'RET': ret,
                'VLT': vlt,
                'COR': cor,
                'DEVENT': devent,
            }
                
            regression_data_list.append(row)
            if ticker in winner_list:
                regression_data_winner_list.append(row)
            elif ticker in loser_list:
                regression_data_loser_list.append(row)
            

regression_df = pd.DataFrame(regression_data_list)
regression_winner_df = pd.DataFrame(regression_data_winner_list)
regression_loser_df = pd.DataFrame(regression_data_loser_list)

with pd.ExcelWriter('regression_data.xlsx') as excel_writer:
    regression_df.to_excel(excel_writer=excel_writer, index=False)

with pd.ExcelWriter('regression_data_winner.xlsx') as excel_writer:
    regression_winner_df.to_excel(excel_writer=excel_writer, index=False)

with pd.ExcelWriter('regression_data_loser.xlsx') as excel_writer:
    regression_loser_df.to_excel(excel_writer=excel_writer, index=False)
